aufgabenzettel01 changes

+++ Aufgabenzettel01/src/TemperatureConverter.java
@@ -3,7 +3,7 @@ import Prog1Tools.IOTools;
 class TemperatureConverter {
 
     public static void main(String[] args) {
-        int temperatureInCelsius = IOTools.readInt("Temperaturwert in Celsius eingeben: ");
+        double temperatureInCelsius = IOTools.readDouble("Temperaturwert in Celsius eingeben: ");
         System.out.println(temperatureInCelsius * 1.8 + 32);
     }
 }
